requirements.txt
flask==3.0.0
flask-cors==4.0.0
twilio==9.0.0
python-dotenv==1.0.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
openai==1.3.0
requests==2.31.0
google-auth==2.26.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.108.0
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.7
python-dateutil==2.8.2
flask-jwt-extended==4.5.2
webrtcvad==2.0.10
sounddevice==0.4.6
urllib3>=2.0.0
numpy<2.0.0,>=1.26.0
